refactored and cleanup of JavaProcess.java and WrappingProcess.java (just one step)
